\name{getData}
\alias{getData}
\title{Function to get data}
\usage{
getData(rdaName)
}
\arguments{
  \item{rdaName}{name of the rda to retrieve}
}
\value{
data.frame this funtion returns a data frame.
}
\description{
This function simply calls the datasets stored with the R
package and sets up a dataframe to be used for all
susbequent analyses
}

