\name{getDataDefinition}
\alias{getDataDefinition}
\title{Function to return Data Definition Table}
\usage{
getDataDefinition(hkmData)
}
\arguments{
  \item{hkmData}{Source dataframe for all data included for
  consideration in Hollister et al.}
}
\description{
This function reads in the data included with the hkm
package, returns all the names of the variables that were
included and builds a table with variable definitions and
source.
}

